% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_header.R
\name{apply_header_names}
\alias{apply_header_names}
\title{Standardize CDEC header}
\usage{
apply_header_names(df)
}
\arguments{
\item{df}{A data frame returned by cdec.R}
}
\value{

}
\description{
Standardize CDEC header
}
