% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdec.R
\name{cdec}
\alias{cdec}
\title{Extract CDEC Data}
\usage{
cdec(station, sensor, duration, start_date = today() - 7,
  end_date = today() - 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Extract CDEC Data
}
\examples{
cdec('SHA', '15', 'D', '2018-01-01', '2018-02-01')
}
