% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builder.R
\name{url_builder}
\alias{url_builder}
\title{Build URL for CDEC}
\usage{
url_builder(station, sensor, duration, start_date, end_date, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Build URL for CDEC
}
\examples{
url_builder('SHA', '15', 'D', '2018-01-01', '2018-02-01')
}
